Move PR stuff into contracts
