+++ source/Calamari/Kubernetes/SpecialVariables.cs
@@ -183,18 +183,6 @@ namespace Calamari.Kubernetes
                     public const string Sources = "sources";
                 }
             }
-            
-            public static class PullRequestCreated
-            {
-                public const string Name = "pull-request-created";
-
-                public static class Attributes
-                {
-                    public const string PullRequestUri = "pullRequestUri";
-                    public const string RepositoryUri = "repositoryUri";
-                    public const string VendorName = "vendorName";
-                }
-            }
         }
     }
 }
